requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Authlib==1.3.2
requests==2.32.3
python-dotenv==1.0.1
